fixed and moved testFixtures
SortingParam.ID -> SortingParam.Price
